Button Class Created, Board and Cell classes edited
Since the previous board and cell classes were from the homework helper I changed a couple of things. I created a game driver class for us to potentially use (not necessary but could be helpful). I also made a way for the buttons to be round to help with the board creation.