progress
I got the computer to make moves at random if it does not try to evaluate for the opponent having 3 in a row. every time  the computer evaluates the board it crashes on the first cell saying its null. its a work in progress but ill look at it with fresh eyes tomorrow